functions/extract-newspaper-titles/requirements.txt
functions-framework==3.*
pandas==2.3.3
feedparser==6.0.12
google-cloud-storage==3.6.0
google-api-core==2.28.1
google-auth==2.43.0
google-resumable-media==2.8.0
proto-plus==1.26.1
protobuf==6.33.1
google-crc32c==1.7.1